+++ functions/cp-angular.js
@@ -1,0 +1,9 @@
+const fs = require("fs-extra");
+
+(async() => {
+  const src = "../dist";
+  const copy = "./dist";
+
+  await fs.remove(copy);
+  await fs.copy(src, copy);
+})();