+++ src/app/pages/account-page/account-page.component.ts
@@ -1,13 +1,22 @@
-import {Component} from "@angular/core";
+import {Component, OnInit} from "@angular/core";
 
 import {AuthService} from "../../services/auth.service";
+import {SeoService} from "../../services/seo.service";
 
 @Component({
   selector: "app-account-page",
   templateUrl: "./account-page.component.html",
   styleUrls: ["./account-page.component.scss"]
 })
-export class AccountPageComponent{
-  constructor(public auth: AuthService){
+export class AccountPageComponent implements OnInit{
+  constructor(public auth: AuthService,
+              private seo: SeoService){
+  }
+
+  ngOnInit(){
+    this.seo.generateTags({
+      title: "Account - BZList",
+      description: "Manage your account"
+    });
   }
 }